Reimplement autocomplete element
The third-party paper-autocomplete element LOOT was using hasn't been
updated to support Polymer 3 yet, so this reimplements it by wrapping
a React autosuggest element.

This might be more work than porting the Polymer element, but I want to
pivot away from Polymer towards React, so this was a good experiment
in doing that.